HHuginn: Fix (re)assign operator compilation.
This commit fixes two issues:

Introduction `+=' operator introduced a bug
which manifested during compilation of following code:

k = 1;
k = 1.0;

That code did not compile due to congruency that was
enforced on `=' too, instead of only on `+='.

Second issue is that reassignment `+=' operator
marked identifier as /written/ instead of as /read/.
Marking code was in OCompiler::defer_make_variable()
and now it is in OCompiler::dispatch_assign().